requirements.txt
# Core dependencies for Adaptive Traffic Signal Control System
# Compatible with Python 3.13 and Windows

# Numerical and scientific computing
numpy>=2.0.0
scipy>=1.10.0
pandas>=2.0.0

# Machine Learning and Deep Learning
tensorflow>=2.20.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
Pillow>=10.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Data validation and configuration
pydantic>=2.8.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Hyperparameter optimization
optuna>=3.0.0

# Logging and monitoring
tensorboard>=2.20.0
wandb>=0.15.0

# Traffic simulation
eclipse-sumo>=1.15.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Performance and optimization
numba>=0.57.0
joblib>=1.3.0

# Network and I/O
requests>=2.31.0
aiohttp>=3.8.0

# Environment management
python-dotenv>=1.0.0
